feat: Add filter input to DataTableToolbar component
This commit adds a new filter input field to the DataTableToolbar component in order to allow users to filter the data table based on a specific column. The filter input is identified by the 'filter' id and is used to update the filter value for the corresponding column. This enhancement improves the usability and functionality of the data table.